prj/itrlm_rag_project/requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
nltk==3.9
hydra-core==1.3.2
PyYAML==6.0.1
tqdm==4.66.4

# PyTorch (Apple Silicon / MPS)
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# Hugging Face / NLP
transformers==4.41.1
accelerate==0.30.1
sentence-transformers==2.6.1
datasets==2.19.0
evaluate==0.4.1
huggingface-hub==0.23.0

# Retrieval / Ranking
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Utils
regex==2024.5.15
langdetect==1.0.9
deep-translator==1.11.4

# FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
python-multipart==0.0.9
